Upgraded Match authoring to Angular. #2893
