+++ src/main/webapp/wise5/components/match/match-authoring/match-authoring.component.html
@@ -1,0 +1,221 @@
+<mat-form-field class="prompt">
+  <mat-label i18n>Prompt</mat-label>
+  <textarea matInput
+      [ngModel]="authoringComponentContent.prompt"
+      (ngModelChange)="promptChanged($event)"
+      placeholder="Enter Prompt Here"
+      i18n-placeholder
+      cdkTextareaAutosize>
+</textarea>
+</mat-form-field>
+<div class="section">
+  <span i18n>Choices</span>
+  <button mat-raised-button
+      color="primary"
+      class="authoring-button add-button"
+      (click)="addChoice()"
+      matTooltip="Add Choice"
+      matTooltipPosition="above"
+      i18n-matTooltip
+      aria-label="Add"
+      i18n-aria-label>
+    <mat-icon>add</mat-icon>
+  </button>
+  <div *ngIf="authoringComponentContent.choices.length === 0" i18n>
+      There are no choices. Click the "Add Choice" button to add a choice.
+  </div>
+  <div *ngFor="let choice of authoringComponentContent.choices; index as choiceIndex; first as isFirst; last as isLast"
+       fxLayout="row" fxLayoutAlign="start center">
+    <mat-form-field class="name">
+      <mat-label i18n>Choice Name</mat-label>
+      <input matInput
+          [(ngModel)]="choice.value"
+          (ngModelChange)="inputChange.next($event)"
+          placeholder="Type text or choose an image"
+          i18n-placeholder/>
+    </mat-form-field>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        (click)="chooseChoiceAsset(choice)"
+        matTooltip="Choose an Image"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Choose an Image"
+        i18n-aria-label>
+      <mat-icon>insert_photo</mat-icon>
+    </button>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        [disabled]="isFirst"
+        (click)="moveChoiceUp(choiceIndex)"
+        matTooltip="Move Up"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Up"
+        i18n-aria-label>
+      <mat-icon>arrow_upward</mat-icon>
+    </button>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        [disabled]="isLast"
+        (click)="moveChoiceDown(choiceIndex)"
+        matTooltip="Move Down"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Down"
+        i18n-aria-label>
+      <mat-icon>arrow_downward</mat-icon>
+    </button>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        (click)="deleteChoice(choiceIndex)"
+        matTooltip="Delete Choice"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Delete"
+        i18n-aria-label>
+      <mat-icon>delete</mat-icon>
+    </button>
+  </div>
+</div>
+<div class="section">
+  <mat-form-field class="name">
+    <mat-label i18n>Source Bucket Name</mat-label>
+    <input matInput
+        [(ngModel)]="authoringComponentContent.choicesLabel"
+        (ngModelChange)="inputChange.next($event)"/>
+  </mat-form-field>
+  <br/>
+  <span i18n>Target Buckets</span>
+  <button mat-raised-button
+      color="primary"
+      class="authoring-button add-button"
+      (click)="addBucket()"
+      matTooltip="Add Target Bucket"
+      matTooltipPosition="above"
+      i18n-matTooltip
+      aria-label="Add"
+      i18n-aria-label>
+    <mat-icon>add</mat-icon>
+  </button>
+  <div *ngIf="authoringComponentContent.buckets.length === 0" i18n>
+    There are no target buckets. Click the "Add Target Bucket" button to add a bucket.
+  </div>
+  <div *ngFor="let bucket of authoringComponentContent.buckets; index as bucketIndex; first as isFirst; last as isLast"
+      fxLayout="row" fxLayoutAlign="start center">
+    <mat-form-field class="name">
+      <mat-label i18n>Target Bucket Name</mat-label>
+      <input matInput
+          [(ngModel)]="bucket.value"
+          (ngModelChange)="inputChange.next($event)"
+          placeholder="Type text or choose an image"
+          i18n-placeholder/>
+    </mat-form-field>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        (click)="chooseBucketAsset(bucket)"
+        matTooltip="Choose an Image"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Choose an Image"
+        i18n-aria-label>
+      <mat-icon>insert_photo</mat-icon>
+    </button>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        [disabled]="isFirst"
+        (click)="moveBucketUp(bucketIndex)"
+        matTooltip="Move Up"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Up"
+        i18n-aria-label>
+      <mat-icon>arrow_upward</mat-icon>
+    </button>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        [disabled]="isLast"
+        (click)="moveBucketDown(bucketIndex)"
+        matTooltip="Move Down"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Down"
+        i18n-aria-label>
+      <mat-icon>arrow_downward</mat-icon>
+    </button>
+    <button mat-raised-button
+        color="primary"
+        class="authoring-button"
+        (click)="deleteBucket(bucketIndex)"
+        matTooltip="Delete Bucket"
+        matTooltipPosition="above"
+        i18n-matTooltip
+        aria-label="Delete"
+        i18n-aria-label>
+      <mat-icon>delete</mat-icon>
+    </button>
+  </div>
+</div>
+<div class="section">
+  <div class="choice-ordered-checkbox">
+    <p i18n>Feedback</p>
+    <mat-checkbox
+        color="primary"
+        [(ngModel)]="authoringComponentContent.ordered"
+        (ngModelChange)="componentChanged()"
+        i18n>
+      Choices need to be ordered within buckets
+    </mat-checkbox>
+  </div>
+  <div *ngFor="let bucketFeedback of authoringComponentContent.feedback"
+      class="section">
+    <p i18n>
+      Bucket Name: {{getBucketNameById(bucketFeedback.bucketId)}}
+    </p>
+    <div *ngFor="let choiceFeedback of bucketFeedback.choices"
+        class="choice-feedback-container">
+      <span i18n>
+        Choice: {{getChoiceTextById(choiceFeedback.choiceId)}}
+      </span>
+      <div layout="row">
+        <mat-form-field class="feedback">
+          <mat-label i18n>Feedback</mat-label>
+          <input matInput
+              color="primary"
+              [(ngModel)]="choiceFeedback.feedback"
+              (ngModelChange)="feedbackChange.next($event)"/>
+        </mat-form-field>
+        <mat-checkbox
+            color="primary"
+            [(ngModel)]="choiceFeedback.isCorrect"
+            (ngModelChange)="isCorrectClicked(choiceFeedback)"
+            layout="row"
+            i18n>
+          Is Correct
+        </mat-checkbox>
+      </div>
+      <div *ngIf="authoringComponentContent.ordered && choiceFeedback.isCorrect" fxLayout="row">
+        <mat-form-field class="position">
+          <mat-label i18n>Position</mat-label>
+          <input matInput
+              type="number"
+              [(ngModel)]="choiceFeedback.position"
+              (ngModelChange)="feedbackChange.next($event)"/>
+        </mat-form-field>
+        <mat-form-field class="incorrect-position-feedback">
+          <mat-label i18n>Incorrect Position Feedback</mat-label>
+          <input matInput
+              [(ngModel)]="choiceFeedback.incorrectPositionFeedback"
+              (ngModelChange)="feedbackChange.next($event)"/>
+        </mat-form-field>
+      </div>
+    </div>
+  </div>
+</div>
