+++ src/main/webapp/wise5/components/match/matchAuthoringComponentModule.ts
@@ -1,15 +1,18 @@
 'use strict';
 
 import * as angular from 'angular';
-import { downgradeInjectable } from '@angular/upgrade/static';
+import { downgradeComponent, downgradeInjectable } from '@angular/upgrade/static';
 import { MatchService } from './matchService';
-import MatchAuthoring from './matchAuthoring';
 import { EditMatchAdvancedComponent } from './edit-match-advanced/edit-match-advanced.component';
+import { MatchAuthoring } from './match-authoring/match-authoring.component';
 
 let matchAuthoringComponentModule = angular
   .module('matchAuthoringComponentModule', ['pascalprecht.translate'])
   .service('MatchService', downgradeInjectable(MatchService))
-  .component('matchAuthoring', MatchAuthoring)
+  .directive(
+    'matchAuthoring',
+    downgradeComponent({ component: MatchAuthoring }) as angular.IDirectiveFactory
+  )
   .component('editMatchAdvanced', EditMatchAdvancedComponent)
   .config([
     '$translatePartialLoaderProvider',
